requirements.txt
pyTelegramBotAPI == 3.7.9
datetime
pytz == 2021.1
mysql-connector-python 8.0.25